Fish will now automatically spawn and proceed to swim around bobber - Need to work on having the fish appear only after one has become interested
